refactor(examples): separate additional example UI (#3186)
